Moved builde of opt system directly to services
This is done in order to specify interaction between view and GraphService. Maybe GraphService is going to be constructed and managed as a proxy for builder or builder will directly interact with controller.
