+++ optical_tracer.py
@@ -1168,166 +1168,6 @@ class OpticalSystem(IOpticalSystem):
             self._append_to_beam(initial_vector=initial_vector, node_vector=current_vector)
 
 
-class IOpticalSystemBuilder(ABC):
-    """Interface for any OptSys builder"""
-
-    @property
-    @abstractmethod
-    def optical_system(self):
-        ...
-
-    @property
-    @abstractmethod
-    def vectors(self):
-        ...
-
-    @abstractmethod
-    def reset(self):
-        """Clear all components from optical system"""
-        ...
-
-    @abstractmethod
-    def trace(self, *, vector: Vector):
-        """Wrapping around optical_system.trace()"""
-        ...
-
-    @abstractmethod
-    def add_components(self, components: Union[OpticalComponent, Iterable[OpticalComponent]]):
-        """Appends given optical component or components to the system"""
-        ...
-
-    @abstractmethod
-    def create_component(self, *, name: str, dimensions: Tuple, layers: Iterable[Layer],
-                         material: Material) -> OpticalComponent:
-        ...
-
-    @abstractmethod
-    def create_layer(self, *, name: str, boundary: Callable, side: Side) -> Layer:
-        ...
-
-    @abstractmethod
-    def create_material(self, *, name: str, transmittance: float, refractive_index: float) -> Material:
-        ...
-
-    @abstractmethod
-    def create_vector(self, *, initial_point: Point, lum: float, w_length: float, theta: float, psi: float) -> Vector:
-        ...
-
-    @abstractmethod
-    def create_point(self, *args, **kwargs):
-        ...
-
-
-class OpticalSystemBuilder(IOpticalSystemBuilder):
-    """The concrete builder of optical systems"""
-
-    def __init__(self):
-        self._optical_system = None
-        self._vectors = []
-
-    @property
-    def optical_system(self):
-        if not isinstance(self._optical_system, OpticalSystem):
-            raise UnspecifiedFieldException(f'Optical system currently hasn''t been initialised properly. '
-                                            'Use builder.reset() to create a new optical system.')
-        return self._optical_system
-
-    @optical_system.setter
-    def optical_system(self, obj: OpticalSystem):
-        if not isinstance(obj, OpticalSystem):
-            raise UnspecifiedFieldException(f'Wrong argument type. '
-                                            f'Supposed to be OpticalSystem, but was given: {type(obj)}')
-        self._optical_system = obj
-
-    @property
-    def vectors(self):
-        return self._vectors
-
-    @vectors.setter
-    def vectors(self, vectors: Union[Vector, Iterable[Vector]]):
-        if isinstance(vectors, Vector):
-            self._vectors.append(vectors)
-        elif isinstance(vectors, Iterable):
-            self._vectors.extend(vectors)
-        else:
-            raise UnspecifiedFieldException(f'Wrong argument type. '
-                                            f'Supposed to be Vector or iterable of Vectors, but was given: {type(vectors)}')
-
-    def reset(self, *, default_medium: Material = OpticalSystem.DEFAULT_CLS_MEDIUM):
-        if not isinstance(default_medium, Material):
-            raise UnspecifiedFieldException(f'Wrong argument type. '
-                                            f'Supposed to be Material, but was given: {type(default_medium)}')
-        new_opt_sys = OpticalSystem(default_medium=default_medium)
-        self.optical_system = new_opt_sys
-        self.vectors = list()
-
-    def trace(self, *, vectors: Union[Vector, Iterable[Vector]]):
-        opt_sys = self.optical_system
-        if isinstance(vectors, Vector):
-            opt_sys.trace(vector=vectors)
-        elif isinstance(vectors, Iterable):
-            for v in vectors:
-                opt_sys.trace(vector=v)
-        else:
-            raise UnspecifiedFieldException(f'Wrong argument type. '
-                                            f'Supposed to be Vector or iterable of Vectors, but was given: {type(vectors)}')
-
-    def trace_all(self):
-        opt_sys = self.optical_system
-        self.trace(vectors=self.vectors)
-
-    def add_components(self, components: Union[OpticalComponent, Iterable[OpticalComponent]]):
-        """Checks if adding is ok and adds component to optical_system.components"""
-        if isinstance(components, OpticalComponent):
-            self.optical_system.add_component(component=components)
-        elif isinstance(components, Iterable):
-            for c in components:
-                self.optical_system.add_component(component=c)
-        else:
-            raise UnspecifiedFieldException(f'Wrong argument type. '
-                                            f'Supposed to be OpticalComponent or iterable of OpticalComponents, '
-                                            f'but was given: {type(components)}')
-
-    def create_component(self, *, name, dimensions=OPT_SYS_DIMENSIONS, layers, material) -> OpticalComponent:
-
-        # TODO. make this aggregation through interface
-        # TODO: make creation through builder
-        new_component = OpticalComponent(name=name, dimensions=dimensions)
-        new_component.material = material
-        for l in layers:
-            new_component.add_layer(layer=l)
-        return new_component
-
-    def create_layer(self, *, name: str, boundary: Callable, side: Side) -> Layer:
-        new_layer = Layer(name=name, boundary=boundary, side=side)
-        return new_layer
-
-    def create_material(self, *, name: str, transmittance: float, refractive_index: float) -> Material:
-        new_material = Material(name=name, transmittance=transmittance, refractive_index=refractive_index)
-        return new_material
-
-    def create_vector(self, *, initial_point: Point, lum: float, w_length: float, theta: float, psi: float) -> Vector:
-        new_vector = Vector(initial_point=initial_point, lum=lum, w_length=w_length, theta=theta, psi=psi)
-        return new_vector
-
-    def create_point(self, *args, **kwargs):
-        """
-        Gets the first tuple of floats of len==3 in positional args: (x: float, y: float, z:float),
-        or keyword arguments for each dimension and returns Point cls object
-        """
-        for arg in args:
-            if len(arg) == 3 and all((isinstance(el, (int, float)) for el in arg)):
-                coords = {'x': arg[0], 'y': arg[1], 'z': arg[2]}
-                return Point(**coords)
-
-        if all((dim in kwargs for dim in ['x', 'y', 'z'])):
-            coords = {dim: kwargs[dim] for dim in ['x', 'y', 'z']}  # filter kwargs with certain keys
-            return Point(**coords)
-
-        raise UnspecifiedFieldException('Wrong arguments. Supposed to be Tuple[float, float, float] in args or '
-                                        '{''x'': float, ''y'': float, ''z'': float} in kwargs,  but was given')
-
-
 def main():
     def create_first_medium():
         first_left_bound = Layer(boundary=lambda y: 0 - y ** 2 / 400, side=Side.LEFT, name='First-left bound')  #
